Got new provider working
